umut3rc yardımlarıyla

+++ main/keys_action.c
@@ -1,6 +1,6 @@
 #include "../header/cub3d.h"
 
-void	keys_action_w(int x, int y, t_game *game)
+void	keys_action_w(double x, double y, t_game *game)
 {
 	if (game->map[(int)((y - sin(game->angle * (M_PI / 180))) / (double)game->img_pixel)][(int)(x / (double)game->img_pixel)] != '1')
 		game->map_player_y -= sin(game->angle * (M_PI / 180));
@@ -35,7 +35,7 @@ void	keys_action_d(int x, int y, t_game *game)
 void	keys_action(int x, int y, t_game *game)
 {
 	if (game->key_w)
-		keys_action_w(x, y, game);
+		keys_action_w((double)x, (double)y, game);
 	if (game->key_a)
 		keys_action_a(x, y, game);
 	if (game->key_s)
